Fixing compile warnings

+++ hci-flashcard-iPhone/hci-flashcard-iPhone/CreateEditListViewController.m
@@ -31,7 +31,7 @@ NSInteger deletionRow = -1;
     
     Deck *deck = self.decks[indexPath.row];
     cell.textLabel.text = deck.name;
-    cell.detailTextLabel.text = [NSString stringWithFormat:@"Cards: %d", deck.cards.count];
+    cell.detailTextLabel.text = [NSString stringWithFormat:@"Cards: %lu", (unsigned long)deck.cards.count];
     
     return cell;
